Backend: Create more legacy text formatting tests
